Вытаскивает логику добавления в историю в отдельный «сервис»
